+++ designer/client/src/components/DndTypes.tsx
@@ -1,0 +1,4 @@
+export const DndTypes = {
+    ELEMENT: "element",
+    VALUE: "value",
+};
