+++ designer/client/src/components/graph/ProcessGraph.tsx
@@ -26,8 +26,8 @@ import { fetchScenarios, getScenariosNames } from "../../reducers/scenarios";
 import { getLayout, getProcessCounts, getScenario } from "../../reducers/selectors/graph";
 import type { Capabilities } from "../../reducers/selectors/other";
 import type { NodeType } from "../../types";
+import { DndTypes } from "../DndTypes";
 import type { Scenario } from "../Process/types";
-import { DndTypes } from "../toolbars/creator/Tool";
 import { jsonToFileInFormData } from "./createFragment";
 import { RECT_HEIGHT, RECT_WIDTH } from "./EspNode/esp";
 import type { Graph } from "./Graph";
